# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental Management System API",
        "description": "A comprehensive car rental management system with user authentication, vehicle management, booking system, payment processing, and review functionality. This system supports both car rentals and property (home) rentals.",
        "contact": {
            "name": "Car Rental System Support",
            "email": "support@carrentalsystem.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Car Rental API Server"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user and send OTP",
                "description": "Register a new user account. The system will send an OTP to both phone and email for verification.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "middle_name",
                                    "last_name",
                                    "email",
                                    "phone",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "User's first name",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "middle_name": {
                                        "description": "User's middle name",
                                        "type": "string",
                                        "example": "A."
                                    },
                                    "last_name": {
                                        "description": "User's last name",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "description": "User's email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "description": "User's phone number (Ethiopian format)",
                                        "type": "string",
                                        "pattern": "^(09|07)\\d{8}$",
                                        "example": "0912345678"
                                    },
                                    "password": {
                                        "description": "User's password (minimum 6 characters)",
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 6,
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "description": "Password confirmation",
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "address": {
                                        "description": "User's address (optional)",
                                        "type": "string",
                                        "example": "Addis Ababa"
                                    },
                                    "city": {
                                        "description": "User's city (optional)",
                                        "type": "string",
                                        "example": "Addis Ababa"
                                    },
                                    "birth_date": {
                                        "description": "User's birth date (optional)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "driver_liscence": {
                                        "description": "Driver's license file (optional, jpg/jpeg/png/pdf)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "digital_id": {
                                        "description": "Digital ID file (optional, jpg/jpeg/png/pdf)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "passport": {
                                        "description": "Passport file (optional, jpg/jpeg/png/pdf)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/verify-phone-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify phone OTP",
                "operationId": "5b23060305a7673f0c43a289efe465b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "otp"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "0912345678"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phone verified successfully"
                    },
                    "400": {
                        "description": "Invalid or expired OTP"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "description": "Authenticate a user and return access token. Supports both regular users and admin users.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User's email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "User"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123xyz..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Account not verified or 2FA required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Phone number not verified. Please enter the OTP sent to your phone."
                                        },
                                        "two_factor_pending": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "email",
                    "phone"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "middle_name": {
                        "type": "string",
                        "example": "A."
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "0912345678"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main Street"
                    },
                    "city": {
                        "type": "string",
                        "example": "Addis Ababa"
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "User",
                            "Admin",
                            "SuperAdmin"
                        ],
                        "example": "User"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Pending",
                            "Active",
                            "Inactive",
                            "Banned"
                        ],
                        "example": "Active"
                    },
                    "is_verified": {
                        "type": "boolean",
                        "example": true
                    },
                    "two_factor_enabled": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "An error occurred"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field": [
                                "The field is required."
                            ]
                        }
                    },
                    "code": {
                        "type": "integer",
                        "example": 400
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}